add methods for clear and default DB tables add pagination logic to repo and VM fix history store add separated events for points and ondrag

+++ app/src/main/java/org/zinchenkodev/antientanimation/data/converters.kt
@@ -16,5 +16,5 @@ fun Line.toEntity(): LineEntity = LineEntity(
 fun LineEntity.toDomain(): Line = Line(
     startDrawing = IntOffset(this.startX, this.startY),
     endDrawing = IntOffset(this.endX, this.endY),
-    color = Color(this.color)
+    color = Color(this.color.toULong())
 )
